+++ app/src/main/java/com/project/camera/AddPerson.java
@@ -47,6 +47,7 @@ public class AddPerson extends AppCompatActivity {
     private TextInputLayout phoneLayout;
     private Button addButton;
     private Bundle extras;
+    private boolean added = false;
 
     TextWatcher nameWatcher = new TextWatcher() {
         @Override
@@ -137,10 +138,10 @@ public class AddPerson extends AppCompatActivity {
             if (validate(name, true) || validate(phone, false)) {
                 Toast toast = Toast.makeText(context, R.string.input_warning, duration);
                 toast.show();
-            } else if (currentPhotoPath == null) {
-                Toast toast = Toast.makeText(context, R.string.image_warning, duration);
-                toast.show();
             } else {
+                if (currentPhotoPath == null){
+                    currentPhotoPath = "";
+                }
                 SecureRandom random = new SecureRandom();
                 int cryptoId = random.nextInt();
                 while (database.checkForCollision(cryptoId, false) || cryptoId == -1) {
@@ -149,6 +150,7 @@ public class AddPerson extends AppCompatActivity {
                 if (database.insertPerson(cryptoId, name, phone, currentPhotoPath)) {
                     Toast toast = Toast.makeText(context, R.string.add_success, duration);
                     toast.show();
+                    added = true;
                 }
                 finish();
             }
@@ -156,6 +158,7 @@ public class AddPerson extends AppCompatActivity {
     }
 
     private void editSetup() {
+        added = true;
         addButton.setText(R.string.update);
         String currentName = extras.getString("Name");
         String currentPhone = extras.getString("Phone");
@@ -163,7 +166,13 @@ public class AddPerson extends AppCompatActivity {
         currentPhotoPath = currentImage;
         nameEditText.setText(currentName);
         phoneEditText.setText(currentPhone);
-        String formatted = getString(R.string.image_path1, currentImage);
+        String formatted;
+        if (Objects.equals(currentImage, "")){
+            formatted = getString(R.string.image_path2);
+        } else {
+            formatted = getString(R.string.image_path1, currentImage);
+        }
+
         imagePath.setText(formatted);
 
         addButton.setOnClickListener(v -> {
@@ -176,11 +185,8 @@ public class AddPerson extends AppCompatActivity {
             if (validate(name, true) || validate(phone, false)) {
                 Toast toast = Toast.makeText(context, R.string.input_warning, duration);
                 toast.show();
-            } else if (currentPhotoPath == null) {
-                Toast toast = Toast.makeText(context, R.string.image_warning, duration);
-                toast.show();
             } else {
-                if(!(currentPhotoPath.equals(currentImage))){
+                if(!(currentPhotoPath.equals(currentImage)) && !Objects.equals(currentImage, "")){
                     File oldImage = new File(currentImage);
                     oldImage.delete();
                 }
@@ -292,6 +298,17 @@ public class AddPerson extends AppCompatActivity {
             return matcher.find() || text.length() <= 1 || text.length() > 31;
         } else {
             return !TextUtils.isDigitsOnly(text) || text.length() != 10;
+        }
+    }
+
+    @Override
+    protected void onStop (){
+        super.onStop();
+        if (currentPhotoPath != null && !added){
+            File pic = new File(currentPhotoPath);
+            if (pic.exists()){
+                pic.delete();
+            }
         }
     }
 }
